+++ apps/skills-showcase/src/__mocks__/next-script.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Script(props: { src?: string; [key: string]: unknown }) {
-  return <script {...props} />;
-}