+++ tests/vitest.config.ts
@@ -16,20 +16,6 @@ export default defineConfig({
           testTimeout: 300_000,
         },
       },
-      {
-        test: {
-          name: "layer3",
-          include: ["layer3/**/*.test.ts"],
-          testTimeout: 360_000,
-        },
-      },
-      {
-        test: {
-          name: "layer4",
-          include: ["layer4/**/*.test.ts"],
-          testTimeout: 3_600_000,
-        },
-      },
     ],
   },
 });